Feature/orders service

+++ composeApp/src/commonMain/kotlin/org/ninjaneers/motager/dashboard/presentation/customers/domain/Customer.kt
@@ -3,6 +3,6 @@ package org.ninjaneers.motager.dashboard.presentation.customers.domain
 data class Customer(
     val id: Int,
     val email: String,
-    val totalPayment: Int,
+    val totalPayment: Double,
     val ordersCount: Int
 )
